Revert "Merge pull request #102 from alexanderwanyoike/refactor/perform-card-38-split-near-limit-files"
This reverts commit d7d04a97f0537c1a1241c7a0db16484076fdfa74, reversing
changes made to 1d0d61e18ea0607b8566dbae9078c6b93d6d1b11.
