+++ crates/vibez-ui/src/app/mod.rs
@@ -150,10 +150,8 @@ mod section_record;
 mod update;
 mod update_media;
 mod update_policy;
-mod update_project;
 mod update_remote;
 mod update_timeline;
-mod update_view;
 mod views_arrangement;
 mod views_automation;
 mod views_browser;
